pyproject.toml
[project]
name = "yolo-inference-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn>=0.31.0",
	"python-multipart>=0.0.12",
	"Pillow>=11.0.0",
	"pydantic>=2.9.0",
	"ultralytics>=8.3.0",
]

[dependency-groups]
dev = [
	"pytest>=8.3.0",
	"pytest-asyncio>=0.24.0",
	"httpx>=0.27.0",
]
